refactor: directory for submodule
