+++ internal/infrastructure/grpc/internal/handler/public/tenant.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/abyssparanoia/rapid-go/internal/infrastructure/grpc/internal/marshaller"
-	public_apiv1 "github.com/abyssparanoia/rapid-go/internal/infrastructure/grpc/pb/rapid/public_api/v1"
 	"github.com/abyssparanoia/rapid-go/internal/usecase/input"
+	public_apiv1 "github.com/abyssparanoia/rapid-go/schema/proto/pb/rapid/public_api/v1"
 )
 
 func (h *PublicHandler) GetTenant(ctx context.Context, req *public_apiv1.GetTenantRequest) (*public_apiv1.GetTenantResponse, error) {
