+++ internal/infrastructure/http/run.go
@@ -18,11 +18,11 @@ import (
 	"github.com/abyssparanoia/rapid-go/internal/infrastructure/dependency"
 	"github.com/abyssparanoia/rapid-go/internal/infrastructure/environment"
 	internal_grpc "github.com/abyssparanoia/rapid-go/internal/infrastructure/grpc"
-	admin_apiv1 "github.com/abyssparanoia/rapid-go/internal/infrastructure/grpc/pb/rapid/admin_api/v1"
-	debug_apiv1 "github.com/abyssparanoia/rapid-go/internal/infrastructure/grpc/pb/rapid/debug_api/v1"
-	public_apiv1 "github.com/abyssparanoia/rapid-go/internal/infrastructure/grpc/pb/rapid/public_api/v1"
 	"github.com/abyssparanoia/rapid-go/internal/infrastructure/http/internal/handler"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/logger"
+	admin_apiv1 "github.com/abyssparanoia/rapid-go/schema/proto/pb/rapid/admin_api/v1"
+	debug_apiv1 "github.com/abyssparanoia/rapid-go/schema/proto/pb/rapid/debug_api/v1"
+	public_apiv1 "github.com/abyssparanoia/rapid-go/schema/proto/pb/rapid/public_api/v1"
 	"github.com/caarlos0/env/v6"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 )
